Add files via upload
I loaded a file with a test of non linear dynamic analysis of the Bathe's beam with Drucker Prager model. The response is wrong. I think it is because possibly the dynamic solver has a error in Newton Raphson algorithm. The displacements is accumulating and don't let oscillation behavior. When elastic model is used the response is correct. The same example on FemLab is performing.